Conflicts:
	nakeduml-runtime/pom.xml

+++ nakeduml-runtime/nakeduml-runtime-tinker-domain/src/main/java/org/nakeduml/runtime/domain/TinkerAuditNode.java
@@ -2,7 +2,7 @@ package org.nakeduml.runtime.domain;
 
 
 public interface TinkerAuditNode extends TinkerNode {
-	int getTransactionNo();
+	Long getTransactionNo();
 //	List<? extends TinkerAuditNode> getNextAuditEntries();
 	TinkerAuditNode getNextAuditEntry();
 	TinkerAuditableNode getOriginal();
